+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 0.12.7
 
+* `Repos`:
+  * `Cache`:
+    * Force `WriteCRUDCacheRepo` to subscribe on new and updated objects of parent repo
+* `Pagination`:
+  * New function `changeResultsUnchecked(Pagination)`
+
 ## 0.12.6
 
 * `MimeeTypes>`: